+++ GLMakie/src/precompiles.jl
@@ -21,7 +21,7 @@ let
             close(screen)
             destroy!(screen)
 
-            include(Makie.SHARED_PRECOMPILE_PATH)
+            foreach(include, Makie.shared_precompile_paths())
             try
                 display(plot(x); visible = false)
             catch